% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectdata.R
\docType{data}
\name{A}
\alias{A}
\alias{A_b_bp}
\alias{A_b}
\title{Bias correction matrices}
\format{A list with entries N512, N256, N128, N64, N32, each containing the bias correction matrix of appropriate size}
\source{
Calculated by hand via /user/s6sebusc/wavelets_verification/general_scripts/Amats_cdtwt.r
}
\usage{
A_b_bp

A_b
}
\description{
Matrices supposedly needed in order to eliminate the effect of "spectral leakage" for the local dtcwt-spectra. Used by biascor( ... ).
}
\examples{
image( A_b_bp$N512 )
}
\keyword{datasets}
