% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{biascor}
\alias{biascor}
\title{spectral bias correction, implemented in FORTRAN}
\usage{
biascor(en, a)
}
\arguments{
\item{en}{\code{ J x nx x ny x 6 } array of squared wavelet coefficients}

\item{a}{bias correction matrix}
}
\value{
an array of the same dimensions as en
}
\description{
Unfold scale and direction into a vector and multiply by A at each grid point
}
\note{
this step can introduce negative values into the spctrum which have no intuitive interpretation in terms of energy and need to be dealt with.
}
