% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagedata.R
\docType{data}
\name{blossom}
\alias{blossom}
\title{Two meteorologists in front of cherry blossoms}
\format{A 512x512 matrix of gray-scale values}
\source{
real life
}
\usage{
blossom
}
\description{
A very beautiful image.
}
\examples{
image(blossom, col=gray.colors(32,0,1))
}
\keyword{datasets}
