% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{boundaries}
\alias{boundaries}
\alias{pad}
\alias{put_in_mirror}
\alias{period_bc}
\title{Various boundary conditions for the 2D wavelet transform.}
\usage{
pad(x, N, Ny = N, value = min(x, na.rm = TRUE))

put_in_mirror(x, N, Ny = N)

period_bc(x, N, Ny = N)
}
\arguments{
\item{x}{a real matrix}

\item{N}{the number of rows of the desired output}

\item{Ny}{the number of columns of the desired output, defaults to N}

\item{value}{the value with which the picture is padded by \code{pad}}
}
\value{
an object of class \code{bc_field}
}
\description{
Various boundary conditions for the 2D wavelet transform.
}
\details{
\code{pad} pads the fields with a constant value on all sides, be careful what you pick here. \code{put_in_mirror} reflects the input at all edges (with repeated end samples), \code{period_bc} simply repeats the input periodically. In any case, you can retrieve the initial area via bc$res[ bc$px, bc$py ].
}
\examples{
bc <- pad( boys, N=300 )
image( bc$res, col=grey.colors(32) )
print( range( bc$res[ bc$px, bc$py ] - boys ) )
}
