% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagedata.R
\docType{data}
\name{boys}
\alias{boys}
\title{Two stromchasers in the sun}
\format{A 256x256 matrix of gray-scale values}
\source{
real life
}
\usage{
boys
}
\description{
Another classic image.
}
\examples{
image(boys, col=gray.colors(32,0,1))
}
\keyword{datasets}
