% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowlevel.R
\name{c2q}
\alias{c2q}
\title{Transform six fields of complex coefficients back into four trees.}
\usage{
c2q(comp)
}
\arguments{
\item{comp}{complex array of dimnesions nx, ny, 6}
}
\value{
a list of low- and high-pass components from the four trees, names LoaHia, LobHib, etc.
}
\description{
This function takes the the six directional complex daughter wavelet coefficients and re-constructs the three combinations of high- and low passes from the four trees (ab, ba, aa, bb).
}
\examples{
c2q( comp )
}
\keyword{drudenfuss}
