% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{dt_analysis}
\alias{dt_analysis}
\title{complete dual-tree analysis}
\usage{
dt_analysis(fld, neg.rm = TRUE, rbr = NULL, pbr = NULL, zbr = NULL,
  mask = NULL, return_fields = TRUE, ...)
}
\arguments{
\item{fld}{a real matrix to be analysed}

\item{rbr, pbr, zbr}{either the number of breaks to use for the histograms of rho, phi and z or vectors containing those breaks.}

\item{mask}{a boolean matrix telling the function which pixels to remove from the analyis (see \code{\link{dt2cen}})}

\item{return_fields}{whether or not you want the function to return the 2D fields of rho, phi, z, u, v and fld itself.}

\item{...}{further parameters passed to \code{\link{fld2dt}}, including wavelet selection and boundary conditions}

\item{neg_rm}{whether or not negative values are immediately removed from the local spectra - recommended.}
}
\value{
and object of class \code{"dtana"}, containing the following: 
\tabular{ll}{ \code{ms}    \tab the mean spectrum \cr
              \code{ce}    \tab the centre of mass of \code{ms} (rho, phi, z)\cr
              \code{rhist} \tab the histogram of radii\cr
              \code{phist} \tab the histogram of angles\cr
              \code{zhist} \tab the histogram of central scales\cr
              \code{rmi}   \tab the bin centres for \code{rhist}\cr
              \code{pmi}   \tab the bin centres for \code{phist}\cr
              \code{zmi}   \tab the bin centres for \code{zhist}
            }
if you set \code{ return_fields=TRUE } (the default), the resulf also contains
\tabular{ll}{ \code{cen}    \tab the output of \code{dt2cen} \cr
              \code{dt}     \tab the output of \code{fld2dt}\cr
              \code{uv}     \tab the centre converted into cathesian coordinates\cr
              \code{fld}    \tab the input field
            }
}
\description{
Transform an input field and calculate a number of summary quantities like the histograms for the three central components, the mean spectrum and the centre of the mean spectrum.
}
\details{
The input is transformed via \code{fld2dt( fld, ... )} and the plugged into \code{dt2cen}. The results are summarized as histograms of radii, angles and central scales. By default, 33 equally spaced breaks between the theoretical extrema of these quantities are used. The default breaks for rho and phi are thus always the same, those for z range from 1 to the largest considered scale.
}
\examples{
dta <- dt_analysis( blossom )
plot( dta )
dtb <- dt_analysis( boys, return_fields=FALSE )
X11()
plot( dtb )
}
\seealso{
\code{\link{fld2dt}}, \code{\link{dt2cen}}
}
