% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{dtmean}
\alias{dtmean}
\title{spatial mean spectrum}
\usage{
dtmean(x)
}
\arguments{
\item{x}{either a J x nx x ny x 6 array of energies (output of dtcwt) or a list of complex wavelet coefficients (the output of \code{dtcwt(...,dec=FALSE)})}
}
\value{
a J x 6 matrix of spatially averaged energies
}
\description{
average the output of fld2dt or dtcwt over space
}
\note{
In the undecimated case, the coefficients are not averaged but summed up and then scaled by the area of the first level. This yields a comparable scale as the undecimated case.
}
