% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{fld2dt}
\alias{fld2dt}
\title{transform a field, handle boundary conditions, return energy}
\usage{
fld2dt(fld, Nx = NULL, Ny = NULL, J = NULL, mode = NULL,
  correct = "b_bp", verbose = FALSE, boundary = "pad")
}
\description{
transform a field, handle boundary conditions, return energy
}
