% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{get_en}
\alias{get_en}
\title{get energy from the dualtree transform}
\usage{
get_en(pyr, correct = "fast", N = ncol(pyr))
}
\arguments{
\item{pyr}{array of \code{ J x nx x ny x 6 } complex wavelet coefficients, output of \code{dtcwt(..., dec=FALSE)}}

\item{correct}{type of correction, either \code{"b"}, \code{"b_bp"} or \code{"fast"}}
}
\value{
an array of the same dimensions as \code{pyr}
}
\description{
square the wavelet coefficients and apply some form of correction
}
\details{
The bias correction matrix should correspond to the filter bank used in the transform. It is computed by brute force for the so-called auto-correlation wavelets, for more details, see Nelson et al 2017 and Eckley et al 2010.
}
\references{
Eckley, Idris A., Guy P. Nason, and Robert L. Treloar. “Locally Stationary Wavelet Fields with Application to the Modelling and Analysis of Image Texture: Modelling and Analysis of Image Texture.” Journal of the Royal Statistical Society: Series C (Applied Statistics), June 21, 2010, no-no. \url{https://doi.org/10.1111/j.1467-9876.2009.00721.x}.

Nelson, J. D. B., A. J. Gibberd, C. Nafornita, and N. Kingsbury. “The Locally Stationary Dual-Tree Complex Wavelet Model.” Statistics and Computing 28, no. 6 (November 2018): 1139–54. \url{https://doi.org/10.1007/s11222-017-9784-0}.
}
\seealso{
\code{\link{A}}
}
