% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idtcwt.R
\name{idtcwt}
\alias{idtcwt}
\title{The 2D inverse dualtree complex wavelet transform}
\usage{
idtcwt(pyr, fb1 = near_sym_b, fb2 = qshift_b, mode = "direct",
  verbose = TRUE, boundaries = "periodic")
}
\description{
The 2D inverse dualtree complex wavelet transform
}
