% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconv.R
\name{my_conv}
\alias{my_conv}
\title{Convolve the columns of a matrix in a varitey of ways}
\usage{
my_conv(mat, fil, dec = TRUE, mode = "direct", odd = FALSE,
  boundaries = "periodic")
}
\arguments{
\item{mat}{a matrix}

\item{fil}{the filter to convolve the columns with}
}
\description{
This function convolves the columns of a matrix mat with a filter fil.
}
\examples{
require( fields )
data( lennon )
my_conv( lennon, c(-1,1) )
}
\keyword{convolution,}
\keyword{wavelets}
