% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterdata.R
\docType{data}
\name{near_sym_b_bp}
\alias{near_sym_b_bp}
\title{A q-shift filter for the second to last levels}
\format{A list of high- and low-pass filters for analysis and synthesis}
\source{
dtcwt python package
}
\usage{
data(qshift_b)
}
\description{
Data from a QTL experiment on gravitropism in
}
\examples{
data(qshift_b)
}
\keyword{datasets}
