% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowlevel.R
\name{q2c}
\alias{q2c}
\title{Transform data from the four trees to six fields of complex coefficients.}
\usage{
q2c(q)
}
\arguments{
\item{q}{a list of wavelet coefficients named LoaHia, LobHib, HiaLoa, ...}
}
\value{
a complex array of size nx, ny, 6
}
\description{
This function takes the four combinations of high- and low passes from the four trees (ab, ba, aa, bb) and re-arranges them into the six directional complex daughter wavelets.
}
\examples{
q2c( q )
}
\keyword{drudenfuss}
