% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{uvplot}
\alias{uvplot}
\title{plot centre as vectors}
\usage{
uvplot(uv, z = NULL, x = NULL, y = NULL, col = "green",
  zcol = gray.colors(32, 0, 1), n = 42, f = 1, code = 0,
  length = 0.05, ...)
}
\arguments{
\item{uv}{either an array of dimension \code{ nx x ny x 2 }, containing the u- and v-component, result of \code{cen2uv} or an object of class \code{"dtana"}}

\item{z}{image to show in the background, defaults to the absolute velocity}

\item{col}{color of the arrows}

\item{zcol}{color scale for the image}

\item{n}{number of arrows in one direction}

\item{f}{factor by which to enlarge the arrows}

\item{code}{determines the type of arrow, passed to \code{arrows()}}

\item{length}{length of the arrowhead in inches, does nothing if \code{code=0}}

\item{...}{further arguments passed to \code{image}}
}
\description{
display the radial and angular component of the spectrum's centre as arrows.
}
\details{
The pivot of the arrows is at the location to which the u- and v-component belong. By default, no arrowhead is displayed (\code{code=0}) since the egdges detcted by the cdtwt have an orientation but no sign (SW and NE are equivalent). The default size of the arrows is such that a 'velocity' of 1 corresponds to 5% of the shorter image side.
}
\examples{
uv <- cen2uv( dt2cen( fld2dt( boys ) ) )
uvplot( uv, z=boys )
}
\seealso{
\code{\link{cen2uv}}
}
