% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagedata.R
\docType{data}
\name{blossom}
\alias{blossom}
\title{Two meteorologists in front of cherry blossoms}
\format{A 256x256 matrix of gray-scale values}
\source{
Armin Blanke, used with permission.
}
\usage{
blossom
}
\description{
A photograph of two meteorologists in front of the famous Japanese cherry trees in Bonn.
}
\examples{
image(blossom, col=gray.colors(128,0,1))
}
\keyword{datasets}
