% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtcwt.R
\name{dtcwt}
\alias{dtcwt}
\title{The 2D forward dualtree complex wavelet transform}
\usage{
dtcwt(mat, fb1 = near_sym_b, fb2 = qshift_b, J = NULL, dec = TRUE,
  mode = NULL, verbose = TRUE, boundaries = "periodic")
}
\arguments{
\item{mat}{the real matrix we wish to transform}

\item{fb1}{A list of analysis filter coefficients for the first level. Currently only near_sym_b and near_sym_b_bp are implemented}

\item{fb2}{A list of analysis filter coefficients for all following levels. Currently only qshift_b and qshift_b_bp are implemented}

\item{J}{number of levels for the decomposition. Defaults to log2( min(Nx,Ny) ) in the decimated case and log2( min(Nx,Ny) ) - 3 otherwise}

\item{dec}{whether or not the decimated transform is desired}

\item{mode}{how to perform the convolutions, either "direct" or "FFT"}
}
\description{
This function performs the dualtree complex wavelet analysis, either with or withour decimation
}
