% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{fld2dt}
\alias{fld2dt}
\title{transform a field, handle boundary conditions, return energy}
\usage{
fld2dt(fld, Nx = NULL, Ny = NULL, J = NULL, mode = NULL,
  correct = NULL, verbose = FALSE, boundary = "pad",
  fb1 = near_sym_b_bp, fb2 = qshift_b_bp)
}
\description{
transform a field, handle boundary conditions, return energy
}
