% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idtcwt.R
\name{idtcwt}
\alias{idtcwt}
\title{The 2D inverse dualtree complex wavelet transform}
\usage{
idtcwt(pyr, fb1 = near_sym_b, fb2 = qshift_b, verbose = TRUE,
  boundaries = "periodic")
}
\arguments{
\item{pyr}{a list containing arrays of complex coefficients for each level of the decomposition, produced by \code{dtcwt( ..., dec=TRUE )}.}

\item{fb1}{the filter bank for the first level}

\item{fb2}{the filter bank for all following levels}

\item{verbose}{if true, the function will say a few words while doing its thing.}

\item{boundaries}{how to handle the boundary conditions, should be the same as for the decomposition.}
}
\value{
a real array of size \code{ 2N x 2M } where \code{ dim( pyr[[1]] ) = (M,N,6) }.
}
\description{
Reconstructs an image from the pyramid of complex directional wavelet coefficients.
}
\details{
This function re-arranges the six complex daughter coefficients back into the four trees, convolves them with the synthesis wavelets and adds everything up to recover an image. For the \code{near_sym_b} and \code{qshift_b} filter banks, this reconstrcution should be basically perfect. In the case of the the \code{b_bp} filters, non-negligible artifacts appear near +-45° edges.
}
\note{
At present, only boundaries="periodic" actually works :(
}
\examples{
py <- dtcwt( boys )
boys_i <- idtcwt( py )
image( boys - boys_i )
}
\references{
Selesnick, I.W., R.G. Baraniuk, and N.C. Kingsbury. “The Dual-Tree Complex Wavelet Transform.” IEEE Signal Processing Magazine 22, no. 6 (November 2005): 123–51. \url{https://doi.org/10.1109/MSP.2005.1550194}.
}
\seealso{
\code{\link{dtcwt}}
}
