% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconv.R
\name{my_conv}
\alias{my_conv}
\title{Column-convolutions}
\usage{
my_conv(mat, fil, dec = TRUE, odd = FALSE, inverse = FALSE)
}
\arguments{
\item{mat}{a matrix}

\item{fil}{the filter to convolve the columns with}

\item{dec}{if \code{TRUE}, every second row is discarded after the convolution}

\item{odd}{if \code{TRUE}, the first row is discarded, otherwise the second row is.}

\item{inverse}{set \code{TRUE} for the inverse transformation}
}
\value{
a matrix with as many columns and either the same (\code{dec=FALSE}) or half (\code{dec=TRUE}) the number of rows as mat
}
\description{
This function convolves the columns of a matrix mat with a filter fil.
}
\details{
This functions does all of the actual computations inside the wavelet transform via FFT. If dec=TRUE or inverse=TRUE, the filters may be longer than the data. In these cases, the data is periodically extended to exceed the filter size.
}
\examples{
dboysdy <- my_conv( blossom, c(-1,1), dec=FALSE )
dboysdx <- t( my_conv( t(blossom), c(-1,1), dec=FALSE ) )
par( mfrow=c(1,2) )
image( dboysdx, col=gray.colors(32) )
image( dboysdy, col=gray.colors(32) )
}
\keyword{convolution,}
\keyword{wavelets}
